+++ R/plots.R
@@ -38,9 +38,9 @@
 #'   time-by-species-by-size array, such as those returned by [getFMort()] or
 #'   [getPredMort()] on a `MizerSim`.
 #'
-#' The same array objects can be passed to [ggplotly()] to produce interactive
-#' versions, for example `ggplotly(getBiomass(sim))` or
-#' `ggplotly(getEncounter(params))`. To add another compatible array to an
+#' The same array objects can be passed to [plotHover()] to produce
+#' hover-enabled plotly versions, for example `plotHover(getBiomass(sim))` or
+#' `plotHover(getEncounter(params))`. To add another compatible array to an
 #' existing ggplot, use [addPlot()]. To compare two compatible mizer arrays
 #' directly, use [plot2()]. To plot cumulative distributions over body size,
 #' use [plotCDF()]. To visualise how spectra or rates change through time, use
@@ -74,7 +74,7 @@
 #' counterparts, for example [plotlyBiomass()] or [plotlySpectra()], for
 #' interactive exploration. Generic and compositional plotting APIs, such as
 #' [plot()], [plot2()], [plotRelative()] and [addPlot()], do not have separate
-#' plotly wrappers. Use [ggplotly()] on the ggplot object they return.
+#' plotly wrappers. Use [plotHover()] on the ggplot object they return.
 #'
 #' While most plot functions take their data from a MizerSim object, some of
 #' those that make plots representing data at a single time can also take their
@@ -104,7 +104,7 @@
 #' # Generic plot methods
 #' plot(sim)
 #' plot(getBiomass(sim), species = c("Cod", "Herring"))
-#' ggplotly(getBiomass(sim))
+#' plotHover(getBiomass(sim))
 #'
 #' # Named plot functions
 #' plotFeedingLevel(sim)
@@ -281,12 +281,34 @@ mizer_tooltip_vars <- function(frame, group_var, x_var, y_var,
     unique(c(tooltip, extra))
 }
 
-#' @exportS3Method plotly::ggplotly
-ggplotly.mizer_plot <- function(p = ggplot2::last_plot(), ...,
-                                tooltip = attr(p, "mizer_tooltip") %||%
-                                    "all") {
-    class(p) <- setdiff(class(p), "mizer_plot")
-    ggplotly(p, ..., tooltip = tooltip)
+#' Create a hover-enabled plotly plot from a mizer object
+#'
+#' Creates an interactive plotly version of a mizer plot. Can be called on any
+#' mizer array object (such as those returned by [getEncounter()],
+#' [getBiomass()], etc.) or on any `mizer_plot` object returned by the named
+#' plot functions such as [plotBiomass()], [plotSpectra()], etc. In the first
+#' case it is a shorthand for `plotly::ggplotly(plot(x, ...))`.
+#'
+#' @param x A `mizer_plot`, `ArraySpeciesBySize`, `ArrayTimeBySpecies`, or
+#'   `ArrayTimeBySpeciesBySize` object.
+#' @param ... Arguments passed to the corresponding `plot()` method for mizer
+#'   array objects, or to [plotly::ggplotly()] for `mizer_plot` objects.
+#' @return A plotly object.
+#' @seealso [plot()], [plotBiomass()], [plotSpectra()], [plotting_functions]
+#' @concept plotting functions
+#' @export
+plotHover <- function(x, ...) UseMethod("plotHover")
+
+#' @rdname plotHover
+#' @param tooltip Character vector of aesthetic names to include in the tooltip,
+#'   or `"all"` for all aesthetics. Defaults to the tooltip stored in the
+#'   `mizer_plot` object.
+#' @export
+plotHover.mizer_plot <- function(x = ggplot2::last_plot(), ...,
+                                 tooltip = attr(x, "mizer_tooltip") %||%
+                                     "all") {
+    class(x) <- setdiff(class(x), "mizer_plot")
+    plotly::ggplotly(x, ..., tooltip = tooltip)
 }
 
 plotComparisonDataFrame <- function(frame1, frame2, params,
@@ -336,22 +358,23 @@ plotComparisonDataFrame <- function(frame1, frame2, params,
     ybreaks <- waiver()
     if (ytrans == "log10") ybreaks <- log_breaks(n = y_ticks)
 
+    frame$LineSpec <- frame[[legend_var]]
     p <- ggplot(frame,
-                aes(group = interaction(.data[[group_var]], .data[["Model"]]))) +
+                aes(group = interaction(.data[[legend_var]], .data[["Model"]]),
+                    colour = .data[[legend_var]])) +
         scale_y_continuous(trans = ytrans, breaks = ybreaks,
                            labels = prettyNum, name = ylab,
                            limits = ylim) +
         scale_x_continuous(trans = xtrans, breaks = xbreaks, name = xlab,
                            limits = xlim) +
         geom_line(aes(x = .data[[x_var]], y = .data[[y_var]],
-                      colour = .data[[legend_var]],
                       linetype = .data[["Model"]],
-                      linewidth = .data[[legend_var]])) +
+                      linewidth = .data[["LineSpec"]])) +
         scale_colour_manual(values = linecolour) +
         scale_linetype_discrete(drop = FALSE) +
         scale_discrete_manual("linewidth", values = linesize)
-    make_mizer_plot(p, mizer_tooltip_vars(frame, group_var, x_var, y_var,
-                                          legend_var, extra = "Model"))
+    make_mizer_plot(p, mizer_tooltip_vars(frame, legend_var, x_var, y_var,
+                                          extra = "Model"))
 }
 
 plotRelativeDataFrame <- function(frame1, frame2, params,
@@ -397,6 +420,7 @@ plotRelativeDataFrame <- function(frame1, frame2, params,
     }
     linecolour <- params@linecolour[legend_levels]
     linesize <- make_linesize(legend_levels, highlight)
+    frame$LineSpec <- frame[[legend_var]]
 
     xbreaks <- waiver()
     if (xtrans == "log10") xbreaks <- log_breaks()
@@ -409,7 +433,7 @@ plotRelativeDataFrame <- function(frame1, frame2, params,
                    colour = "dark grey", linewidth = 0.75) +
         geom_line(aes(x = .data[[x_var]], y = .data[["rel_diff"]],
                       colour = .data[[legend_var]],
-                      linewidth = .data[[legend_var]])) +
+                      linewidth = .data[["LineSpec"]])) +
         scale_colour_manual(values = linecolour) +
         scale_discrete_manual("linewidth", values = linesize)
     make_mizer_plot(p, mizer_tooltip_vars(frame, group_var, x_var, "rel_diff",
@@ -637,7 +661,7 @@ plotlyBiomass <- function(object,
     argg <- as.list(environment())
     argg$start_time <- NULL
     argg$end_time <- NULL
-    ggplotly(do.call("plotBiomass", argg),
+    plotHover(do.call("plotBiomass", argg),
              tooltip = c("Species", "Year", "Biomass"))
 }
 
@@ -783,7 +807,7 @@ plotlyYield <- function(object, sim2,
                         highlight = NULL,
                         ...) {
     argg <- as.list(environment())
-    ggplotly(do.call("plotYield", argg),
+    plotHover(do.call("plotYield", argg),
              tooltip = c("Species", "Year", "Yield"))
 }
 
@@ -897,14 +921,18 @@ plotYieldGear.MizerSim <- function(object,
     # Need to keep species in order for legend
     species_levels <- intersect(names(params@linecolour), ym$Species)
     ym$Species <- factor(ym$Species, levels = species_levels)
+    ym$Legend <- ym$Species
     linesize <- make_linesize(species_levels, highlight)
-    ggplot(ym) +
-        geom_line(aes(x = Year, y = Yield, colour = Species,
-                      linetype = Gear, linewidth = Species)) +
-        scale_y_continuous(trans = "log10", name = "Yield [g]",
-                           limits = ylim) +
-        scale_colour_manual(values = params@linecolour[species_levels]) +
-        scale_discrete_manual("linewidth", values = linesize)
+    make_mizer_plot(
+        ggplot(ym) +
+            geom_line(aes(x = Year, y = Yield, colour = Species,
+                          linetype = Gear, linewidth = Legend)) +
+            scale_y_continuous(trans = "log10", name = "Yield [g]",
+                               limits = ylim) +
+            scale_colour_manual(values = params@linecolour[species_levels]) +
+            scale_discrete_manual("linewidth", values = linesize),
+        c("Species", "Gear", "Year", "Yield")
+    )
 }
 
 #' @rdname plotYieldGear
@@ -916,8 +944,8 @@ plotlyYieldGear <- function(object, species = NULL,
                             ylim = c(NA, NA), tlim = c(NA, NA),
                             highlight = NULL, ...) {
     argg <- as.list(environment())
-    ggplotly(do.call("plotYieldGear", argg),
-             tooltip = c("Species", "Year", "Yield"))
+    plotHover(do.call("plotYieldGear", argg),
+             tooltip = c("Species", "Gear", "Year", "Yield"))
 }
 
 #' Plot abundance and biomass spectra and compare between objects
@@ -1222,6 +1250,7 @@ plot_spectra <- function(params, n, n_pp,
     if (identical(size_axis, "l")) {
         plot_dat <- filter_plot_length_limits(plot_dat, llim)
     }
+    names(plot_dat)[2] <- y_label
     if (return_data) return(plot_dat)
 
     plotDataFrame(plot_dat, params, xlab = plot_size_xlab(size_axis),
@@ -1384,11 +1413,13 @@ plot_cdf <- function(plot_dat, params, power, normalise, log_x, log_y, wlim, lli
     cdf_dat <- prepare_spectra_cdf_data(plot_dat, params,
                                         normalise = normalise)
     cdf_dat <- convert_plot_size_axis(cdf_dat, params, size_axis)
+    cdf_y <- cdf_y_label(power, normalise)
+    names(cdf_dat)[2] <- cdf_y
     if (return_data) return(cdf_dat)
 
     plotDataFrame(cdf_dat, validParams(params),
                   xlab = plot_size_xlab(size_axis),
-                  ylab = cdf_y_label(power, normalise),
+                  ylab = cdf_y,
                   xtrans = if (log_x) "log10" else "identity",
                   ytrans = if (log_y) "log10" else "identity",
                   xlim = plot_size_xlim(wlim, size_axis, llim), ylim = ylim,
@@ -1397,13 +1428,14 @@ plot_cdf <- function(plot_dat, params, power, normalise, log_x, log_y, wlim, lli
 
 prepare_spectra_cdf_data <- function(plot_dat, params, normalise = TRUE) {
     params <- validParams(params)
+    y_var <- names(plot_dat)[2]
     plot_dat <- plot_dat[order(plot_dat$Species, plot_dat$w), ]
-    plot_dat$value <- plot_dat$value * spectra_bin_width(plot_dat$w, params)
-    plot_dat$value <- stats::ave(plot_dat$value, plot_dat$Species,
-                                 FUN = cumsum)
+    plot_dat[[y_var]] <- plot_dat[[y_var]] * spectra_bin_width(plot_dat$w, params)
+    plot_dat[[y_var]] <- stats::ave(plot_dat[[y_var]], plot_dat$Species,
+                                    FUN = cumsum)
     if (normalise) {
-        totals <- stats::ave(plot_dat$value, plot_dat$Species, FUN = max)
-        plot_dat$value <- plot_dat$value / totals
+        totals <- stats::ave(plot_dat[[y_var]], plot_dat$Species, FUN = max)
+        plot_dat[[y_var]] <- plot_dat[[y_var]] / totals
     }
     plot_dat
 }
@@ -1561,9 +1593,7 @@ plotlySpectra2 <- function(object1, object2, name1 = "First",
     if (!missing(power)) {
         args$power <- power
     }
-    ggplotly(do.call("plotSpectra2", args),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = c("value", "Model")))
+    plotHover(do.call("plotSpectra2", args))
 }
 
 #' @rdname plotSpectra
@@ -1626,7 +1656,7 @@ plotlySpectraRelative <- function(object1, object2,
     if (!missing(power)) {
         args$power <- power
     }
-    ggplotly(do.call("plotSpectraRelative", args),
+    plotHover(do.call("plotSpectraRelative", args),
              tooltip = plot_size_tooltip(size_axis, before = "Legend",
                                          after = "rel_diff"))
 }
@@ -1659,9 +1689,7 @@ plotlyCDF <- function(object, species = NULL,
     if (!missing(power)) {
         args$power <- power
     }
-    ggplotly(do.call("plotCDF", args),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = "value"))
+    plotHover(do.call("plotCDF", args))
 }
 
 #' @rdname plotCDF
@@ -1690,9 +1718,7 @@ plotlyCDF2 <- function(object1, object2, name1 = "First", name2 = "Second",
     if (!missing(power)) {
         args$power <- power
     }
-    ggplotly(do.call("plotCDF2", args),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = c("value", "Model")))
+    plotHover(do.call("plotCDF2", args))
 }
 
 #' @rdname plotSpectra
@@ -1710,9 +1736,7 @@ plotlySpectra <- function(object, species = NULL,
                         size_axis = c("w", "l"), ...) {
     size_axis <- plot_size_axis(size_axis)
     argg <- as.list(environment())
-    ggplotly(do.call("plotSpectra", argg),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = "value"))
+    plotHover(do.call("plotSpectra", argg))
 }
 
 #' Plot the feeding level of species by size
@@ -1875,6 +1899,7 @@ plot_feeding_level <- function(params, feed, species, highlight,
         plot_dat <- subset(plot_dat, value > 0)
     }
 
+    names(plot_dat)[2] <- "Feeding level"
     if (return_data) return(plot_dat)
     x_var <- plot_size_x_var(size_axis)
 
@@ -1899,18 +1924,21 @@ plot_feeding_level <- function(params, feed, species, highlight,
     } else {
         p <- ggplot(plot_dat, aes(group = Species))
     }
-    p + geom_line(aes(x = .data[[x_var]], y = value,
-                      colour = Legend, linetype = Legend, linewidth = Legend)) +
-        scale_x_continuous(name = plot_size_xlab(size_axis),
-                           trans = if (log_x) "log10" else "identity",
-                           limits = plot_size_xlim(wlim, size_axis, llim)) +
-        scale_y_continuous(name = "Feeding Level",
-                           trans = if (log_y) "log10" else "identity") +
-        # Feeding level is naturally bounded in [0, 1] on linear scale.
-        coord_cartesian(ylim = if (log_y) NULL else c(0, 1)) +
-        scale_colour_manual(values = params@linecolour[legend_levels]) +
-        scale_linetype_manual(values = params@linetype[legend_levels]) +
-        scale_discrete_manual("linewidth", values = linesize)
+    make_mizer_plot(
+        p + geom_line(aes(x = .data[[x_var]], y = .data[["Feeding level"]],
+                          colour = Legend, linetype = Legend, linewidth = Legend)) +
+            scale_x_continuous(name = plot_size_xlab(size_axis),
+                               trans = if (log_x) "log10" else "identity",
+                               limits = plot_size_xlim(wlim, size_axis, llim)) +
+            scale_y_continuous(name = "Feeding Level",
+                               trans = if (log_y) "log10" else "identity") +
+            # Feeding level is naturally bounded in [0, 1] on linear scale.
+            coord_cartesian(ylim = if (log_y) NULL else c(0, 1)) +
+            scale_colour_manual(values = params@linecolour[legend_levels]) +
+            scale_linetype_manual(values = params@linetype[legend_levels]) +
+            scale_discrete_manual("linewidth", values = linesize),
+        plot_size_tooltip(size_axis, before = "Species", after = "Feeding level")
+    )
 }
 
 #' @rdname plotFeedingLevel
@@ -1927,9 +1955,7 @@ plotlyFeedingLevel <- function(object,
                              log_x = TRUE, log_y = FALSE, log = NULL, ...) {
     size_axis <- plot_size_axis(size_axis)
     argg <- as.list(environment())
-    p <- ggplotly(do.call("plotFeedingLevel", argg),
-                  tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                              after = "value"))
+    p <- plotHover(do.call("plotFeedingLevel", argg))
 
     # When critical feeding level is included, ggplotly creates traces split by the
     # interaction of Species and Type, which produces a very long combined legend.
@@ -2102,9 +2128,7 @@ plotlyPredMort <- function(object, species = NULL,
                            log_x = TRUE, log_y = FALSE, log = NULL, ...) {
     size_axis <- plot_size_axis(size_axis)
     argg <- as.list(environment())
-    ggplotly(do.call("plotPredMort", argg),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = "value"))
+    plotHover(do.call("plotPredMort", argg))
 }
 
 #' Plot total fishing mortality of each species by size
@@ -2199,9 +2223,7 @@ plotlyFMort <- function(object, species = NULL,
                         log_x = TRUE, log_y = FALSE, log = NULL, ...) {
     size_axis <- plot_size_axis(size_axis)
     argg <- as.list(environment())
-    ggplotly(do.call("plotFMort", argg),
-             tooltip = plot_size_tooltip(size_axis, before = "Species",
-                                         after = "value"))
+    plotHover(do.call("plotFMort", argg))
 }
 
 
@@ -2375,22 +2397,24 @@ plot_growth_curves <- function(params, species,
         plot_dat2$Legend <- "von Bertalanffy"
         plot_dat <- rbind(plot_dat, plot_dat2)
     }
-    if (return_data) return(plot_dat)
-
-    p <- ggplot(filter(plot_dat, Legend == "model")) +
-        geom_line(aes(x = Age, y = value,
-                      colour = Species, linetype = Species,
-                      linewidth = Species))
     y_label <- if (percentage)
         "Percent of maximum size"
     else "Size [g]"
-    # Need to keep species in order for legend
+    names(plot_dat)[3] <- y_label
+    if (return_data) return(plot_dat)
+
     legend_levels <-
         intersect(c(dimnames(params@initial_n)$sp,
                     "Background", "Resource", "Total"),
                   plot_dat$Species)
     plot_dat$Species <- factor(plot_dat$Species, levels = legend_levels)
+    plot_dat$LineSpec <- plot_dat$Species
     linesize <- make_linesize(legend_levels, highlight)
+    p <- ggplot(filter(plot_dat, Legend == "model"),
+                aes(group = Species)) +
+        geom_line(aes(x = Age, y = .data[[y_label]],
+                      colour = .data[["LineSpec"]], linetype = .data[["LineSpec"]],
+                      linewidth = .data[["LineSpec"]]))
     p <- p + scale_x_continuous(name = "Age [Years]",
                                 trans = if (log_x) "log10" else "identity") +
         scale_y_continuous(name = y_label,
@@ -2413,7 +2437,7 @@ plot_growth_curves <- function(params, species,
             if ("von Bertalanffy" %in% plot_dat$Legend) {
                 p <- p + geom_line(data = filter(plot_dat,
                                                  Legend == "von Bertalanffy"),
-                                   aes(x = Age, y = value))
+                                   aes(x = Age, y = .data[[y_label]]))
             }
             if (!is.null(size_at_age)) {
                 size_at_age <- filter(size_at_age, species == selected_species)
@@ -2425,7 +2449,7 @@ plot_growth_curves <- function(params, species,
         } else if (species_panel) { # need to add either no panel if no param
             # for VB or create a panel without VB
             p <- ggplot(plot_dat) +
-                geom_line(aes(x = Age, y = value, colour = Legend)) +
+                geom_line(aes(x = Age, y = .data[[y_label]], colour = Legend)) +
                 scale_x_continuous(name = "Age [years]",
                                    trans = if (log_x) "log10" else "identity") +
                 scale_y_continuous(name = "Size [g]",
@@ -2450,7 +2474,7 @@ plot_growth_curves <- function(params, species,
         }
 
     }
-    return(p)
+    return(make_mizer_plot(p, c("Species", "Age", y_label)))
 
 }
 
@@ -2464,8 +2488,7 @@ plotlyGrowthCurves <- function(object, species = NULL,
                                highlight = NULL,
                                log_x = FALSE, log_y = FALSE, log = NULL) {
     argg <- as.list(environment())
-    ggplotly(do.call("plotGrowthCurves", argg),
-             tooltip = c("Species", "Age", "value"))
+    plotHover(do.call("plotGrowthCurves", argg))
 }
 
 
@@ -2600,7 +2623,8 @@ plot_diet <- function(params, n, diet, species, log_x, log_y, wlim, llim,
     if (sum(species) > 1) {
         p <- p + facet_wrap(vars(Predator))
     }
-    p
+    make_mizer_plot(p, plot_size_tooltip(size_axis,
+                                         after = c("Proportion", "Prey")))
 }
 
 #' @rdname plotDiet
@@ -2612,11 +2636,11 @@ plotlyDiet <- function(object, species = NULL,
                        size_axis = c("w", "l"),
                        log_x = TRUE, log_y = FALSE, log = NULL, ...) {
     size_axis <- plot_size_axis(size_axis)
-    ggplotly(plotDiet(object, species = species, wlim = wlim, llim = llim,
-                      log_x = log_x, log_y = log_y, log = log,
-                      size_axis = size_axis, ...),
-             tooltip = plot_size_tooltip(size_axis, before = "Predator",
-                                        after = c("Proportion", "Prey")))
+    plotHover(plotDiet(object, species = species, wlim = wlim, llim = llim,
+                       log_x = log_x, log_y = log_y, log = log,
+                       size_axis = size_axis, ...),
+              tooltip = plot_size_tooltip(size_axis,
+                                         after = c("Proportion", "Prey")))
 }